requirements.txt
accelerate==0.29.2
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
auto_gptq==0.7.1
backoff==2.2.1
bitsandbytes==0.43.1
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.5
datasets==2.18.0
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docarray==0.40.0
docker==7.0.0
ecdsa==0.19.0
evaluate==0.4.1
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fast-sentence-transformers==0.4.1
fastapi==0.110.1
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.2.0
fuzzywuzzy==0.18.0
gekko==1.1.1
googleapis-common-protos==1.63.0
googletrans==3.0.0
greenlet==3.0.3
grpcio==1.57.0
grpcio-health-checking==1.57.0
grpcio-reflection==1.57.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.22.2
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.11.0
itsdangerous==2.2.0
jcloud==0.3
jina==3.25.1
jina-hubble-sdk==0.39.0
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langcodes==3.4.0
langsmith==0.1.54
language_data==1.2.0
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
onnx==1.16.0
onnxruntime-gpu==1.17.1
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-exporter-prometheus==0.41b0
opentelemetry-instrumentation==0.40b0
opentelemetry-instrumentation-aiohttp-client==0.40b0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-grpc==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
opentelemetry-util-http==0.40b0
optimum==1.19.0
orjson==3.10.0
packaging==23.2
pandas==2.2.2
pathspec==0.12.1
peft==0.10.0
pillow==10.3.0
polars==0.20.19
preshed==3.0.9
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pydantic==1.10.15
Pygments==2.17.2
pymongo==4.7.1
python-benedict==0.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fsutil==0.14.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rfc3986==1.5.0
rich==13.7.1
rouge==1.0.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.6.1
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.4.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.2
triton==2.2.0
txtai==7.1.0
typer==0.9.4
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.23.1
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
